LayoutTests/imported/w3c/web-platform-tests/user-timing/measure-l3.any-expected.txt


PASS When the end mark is given and the start is unprovided, the end time of the measure entry should be the end mark's time, the start time should be 0.
FAIL When the start mark is given and the end is unprovided, the start time of the measure entry should be the start mark's time, the end should be now. assert_greater_than_equal: expected a number greater than or equal to 9.000000000000002 but got 9
PASS When start and end mark are both given, the start time and end time of the measure entry should be the the marks' time, repectively
